Refactor common types
