+++ apps/ai-image-generator/backend/src/actions/aiig-erase-and-fill.ts
@@ -3,6 +3,7 @@ import { AppActionCallContext } from '@contentful/node-apps-toolkit';
 import { PlainClientAPI } from 'contentful-management';
 import { OpenAiApiService } from '../services/openaiApiService';
 import * as nodeFetch from 'node-fetch';
+import { AppActionCallResponse, Image } from '../types';
 
 interface AppActionCallParameters {
   prompt: string;
@@ -10,32 +11,10 @@ interface AppActionCallParameters {
   mask: string;
 }
 
-interface Image {
-  url: string;
-  imageType: string;
+export interface ImageEditResult {
+  type: 'ImageEditResult';
+  images: Image[];
 }
-
-interface ActionError {
-  type: string;
-  message: string;
-  details?: Record<string, any>;
-}
-
-// TODO: Create generic versions of success and error
-export interface AppActionCallResponseSuccess {
-  ok: true;
-  data: {
-    type: 'ImageEditResult';
-    images: Image[];
-  };
-}
-
-export interface AppActionCallResponseError {
-  ok: false;
-  errors: ActionError[];
-}
-
-type AppActionCallResponse = AppActionCallResponseSuccess | AppActionCallResponseError;
 
 async function fetchOpenAiApiKey(cma: PlainClientAPI, appInstallationId: string): Promise<string> {
   const appInstallation = await cma.appInstallation.get({ appDefinitionId: appInstallationId });
@@ -50,7 +29,7 @@ async function fetchOpenAiApiKey(cma: PlainClientAPI, appInstallationId: string)
 export const handler = async (
   payload: AppActionCallParameters,
   context: AppActionCallContext
-): Promise<AppActionCallResponse> => {
+): Promise<AppActionCallResponse<ImageEditResult>> => {
   const {
     cma,
     appActionCallContext: { appInstallationId },
