+++ apps/ai-image-generator/backend/src/actions/aiig-generate-image.spec.ts
@@ -8,13 +8,10 @@ import sinon from 'sinon';
 import { OpenAiApiService } from '../services/openaiApiService';
 import OpenAI from 'openai';
 import sinonChai from 'sinon-chai';
-import {
-  AppActionCallResponseError,
-  AppActionCallResponseSuccess,
-  handler,
-} from './aiig-generate-image';
+import { ImageCreationResult, handler } from './aiig-generate-image';
 import { AppInstallationProps, SysLink } from 'contentful-management';
 import { APIError } from 'openai/error';
+import { AppActionCallResponseError, AppActionCallResponseSuccess } from '../types';
 
 chai.use(sinonChai);
 
@@ -49,7 +46,10 @@ describe('aiigGenerateImage.handler', () => {
   });
 
   it('returns the images result', async () => {
-    const result = (await handler(parameters, context)) as AppActionCallResponseSuccess;
+    const result = (await handler(
+      parameters,
+      context
+    )) as AppActionCallResponseSuccess<ImageCreationResult>;
     expect(result).to.have.property('ok', true);
     expect(result.data).to.have.property('type', 'ImageCreationResult');
     expect(result.data.images).to.deep.include({
